feat: 커스텀 GNB 수정 & 유저 데이터 더미로 로그인
1. 유저 데이터 더미 파일을 만들고 shared의 js에서 해당 데이터를 커스텀 GNB의 프로퍼티로 설정
2. 커스텀 GNB의 프로퍼티가 설정되면 해당 유저의 정보로 로그인된 gnb를 나타내고, 따로 설정이 되지 않으면 로그인 버튼이 있는 gnb를 나타냄
