+++ index.html
@@ -24,7 +24,6 @@
       name="twitter:image"
       content="https://images.unsplash.com/photo-1481627834876-b7833e8f5570?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=928&q=80"
     />
-    <script src="/components/custom-gnb.js"></script>
   </head>
   <body>
     <custom-gnb></custom-gnb>
@@ -99,6 +98,7 @@
       </section>
     </main>
     <custom-footer></custom-footer>
+    <script src="/components/GNB/custom-gnb.js"></script>
     <script src="/components/Footer/custom-footer.js"></script>
   </body>
 </html>
